Added extra parameters to MS/MS Peak picker. Now the user can select the level of MSn scans. I changed the functionality of this module. Before, the module was using the parent scan of the fragment scan as starting point to look for a data point in a mass window range.
Now the module use the mass value of the parent ion to look in scans with MS level 1 within the mass and time window range. This help to avoid problems with data that has an intermediate scan (i.e. Zoom scan) between MS1 and MS2 scans.

The ExtendedPeak class now has a method setMostIntenseFragmentScanNumber(). The original idea of this module was to extend detected peaks using MS/MS Peak Picker. The fragment scan number should be respected.

The MascotSearchTask class now can deal with an error webpage from Mascot server.