Added Base Filtering Integer
I added the so called "Base Filtering Integer". I had troubles naming this variable, but it's possible to describe it in a longer manner (which is what I did in the tooltip).

The base filtering integer filters every Nth element of the current scan(s). It's best shown how it works by example.
If you input 3 as the base filtering integer it will filter this list in such a way:
[1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12]
into
[1, 4, 7, 10]

As with the two previous pull requests, any ideas or changes are appreciated. A name change is very forseeable  It kind of bothers me and I can't think of anything more short and to the point.