Make NeutralLossPlot extend EChartViewer.
As the dataset and the xAxisType can no longer be transferred form NLWindow
to NLPlot via arguments of the NLConstructor, the methods setAxisTypes and
addNeutralLossDataSet were added to NLPlot.

As GUIUtils are still implemeted with swing, use setOnAction here for now.

As the non-swing MenuItems don't work with the awt-action handling and with
the scene.control-action handling, I did not see a way of discrimiating
between different meanings of the same EventType (the menu items doing
different things), I used here a local eventHandler too.
-> Eventually "Highlight Neutral Loss" and "Highlight Percursor" rather
belong into the menu of the NLVisualizerWindow or as a button into the toolbar
=> I tried out with one button (in NLVisualizerWindow).

-> TODO: replace swing.JDialog